invalidation mid-solution now functional
before, the function SLIPsim.m was terminated before some of the outputs were defined. This was rectified by initializing all outputs to NaN except discrep, which was initialized to zero. When invalidating mid-solution, discrep is set to NaN. 

Also made it so that it only plots/animates when solution is valid